requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
Django==4.1.9
django-cors-headers==3.14.0
django-phonenumber-field==7.1.0
django-sse==0.4.1
djangorestframework==3.14.0
filelock==3.9.0
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
markdown2==2.4.8
multidict==6.0.4
openai==0.27.7
phonenumberslite==8.13.11
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
toml==0.10.2
tqdm==4.65.0
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.19.0
yarl==1.9.2
zope.event==4.6
zope.interface==6.0
whitenoise==6.5.0